Refactoring Skynet structure.
