+++ src/main/java/com/innowise/skynet/model/Part.java
@@ -1,0 +1,5 @@
+package com.innowise.skynet.model;
+
+public enum Part {
+    HEAD, TORSO, HAND, FEET
+}
